Changed date types from ints to longs

+++ src/main/kotlin/ui/models/Chat.kt
@@ -6,7 +6,7 @@ import org.jetbrains.exposed.dao.IntEntityClass
 import org.jetbrains.exposed.dao.IntIdTable
 
 object Chats : IntIdTable("chat", "ROWID") {
-    val lastReadTimestamp = integer("last_read_message_timestamp")
+    val lastReadTimestamp = long("last_read_message_timestamp")
 }
 
 class Chat(rowid: EntityID<Int>) : IntEntity(rowid) {
