+++ src/main/kotlin/ui/models/ChatMessageJoin.kt
@@ -8,7 +8,7 @@ import org.jetbrains.exposed.dao.IntIdTable
 object ChatMessageJoins : IntIdTable("chat_message_join", "ROWID") {
     val chatID = integer("chat_id")
     val messageID = integer("message_id")
-    val date = integer("message_date")
+    val date = long("message_date")
 }
 
 class ChatMessageJoin(rowid: EntityID<Int>) : IntEntity(rowid) {
